+++ src/Models/DataFiles/PageData/Component.ts
@@ -1,7 +1,7 @@
 import ColumnPosition from "./ColumnPosition";
 import ComponentType from "./ComponentType";
-import Paragraph from "../../Paragraph";
-import ImageFile from "../../ImageFile";
+import ParagraphData from "../../ParagraphData";
+import ImageData from "../../ImageData";
 import ImageSlider from "./ImageSlider";
 
 class Component {
@@ -10,8 +10,8 @@ class Component {
     order: number = 0;
     columnPosition: ColumnPosition = ColumnPosition.Left;
     componentType: ComponentType = ComponentType.Info;
-    paragraphs?: Paragraph[];
-    imageFiles?: ImageFile[];
+    paragraphs?: ParagraphData[];
+    imageFiles?: ImageData[];
     imageSlider?: ImageSlider = new ImageSlider(); 
 
     constructor(component?: {
@@ -20,8 +20,8 @@ class Component {
         order: number,
         columnPosition: ColumnPosition,
         componentType: ComponentType,
-        paragraphs?: Paragraph[],
-        imageFiles?: ImageFile[],
+        paragraphs?: ParagraphData[],
+        imageFiles?: ImageData[],
         imageSlider?: ImageSlider
     }){
         if(component){
@@ -30,8 +30,8 @@ class Component {
             this.order = component.order ? component.order : 0;
             this.columnPosition = component.columnPosition ? component.columnPosition : ColumnPosition.Left;
             this.componentType = component.componentType ? component.componentType : ComponentType.Info;
-            this.paragraphs = component.paragraphs ? component.paragraphs.map(item => new Paragraph(item)) : [];
-            this.imageFiles = component.imageFiles ? component.imageFiles.map(item => new ImageFile(item)) : [];
+            this.paragraphs = component.paragraphs ? component.paragraphs.map(item => new ParagraphData(item)) : [];
+            this.imageFiles = component.imageFiles ? component.imageFiles.map(item => new ImageData(item)) : [];
             this.imageSlider = component.imageSlider ? component.imageSlider : new ImageSlider();
         } else {
             this.active = false;