+++ src/Models/ParagraphData.ts
@@ -1,25 +1,25 @@
 import Alignment from "./DataFiles/PageData/Alignment";
 
-class Paragraph {
+class ParagraphData {
     display: boolean = false;
     header: boolean = false;
     emphasis: boolean = false;
     text: string = "";
     alignment: Alignment = Alignment.Left;
     
-    constructor(paragraph?: {
+    constructor(paragraphData?: {
         display: boolean,
         header: boolean,
         emphasis: boolean,
         text: string,
         alignment: Alignment
     }){
-        if(paragraph){
-            this.display = paragraph.display ? paragraph.display : false;
-            this.header = paragraph.header ? paragraph.header : false;
-            this.emphasis = paragraph.emphasis ? paragraph.emphasis : false;
-            this.text = paragraph.text ? paragraph.text : "";
-            this.alignment = paragraph.alignment ?  paragraph.alignment : Alignment.Left;
+        if(paragraphData){
+            this.display = paragraphData.display ? paragraphData.display : false;
+            this.header = paragraphData.header ? paragraphData.header : false;
+            this.emphasis = paragraphData.emphasis ? paragraphData.emphasis : false;
+            this.text = paragraphData.text ? paragraphData.text : "";
+            this.alignment = paragraphData.alignment ?  paragraphData.alignment : Alignment.Left;
         } else {
             this.display = false;
             this.header = false;
@@ -30,4 +30,4 @@ class Paragraph {
     }
 }
 
-export default Paragraph;
+export default ParagraphData;