+++ src/Processors/ImageProcessor.tsx
@@ -1,15 +1,15 @@
-import Image from "../Components/Content/Images/ImageFiles/Image/Image";
+import Image from "../Components/Content/Images/ImageFiles/SiteImage/SiteImage";
 import Component from '../Models/DataFiles/PageData/Component';
 import ImageSlider from '../Components/Content/Images/ImageSlider/ImageSlider';
-import ImageFile from '../Models/ImageFile';
+import ImageData from '../Models/ImageData';
 
-import classes from '../Components/Content/Images/ImageFiles/Image/Image.module.css';
+import classes from '../Components/Content/Images/ImageFiles/SiteImage/SiteImage.module.css';
 
 class ImageProcessor {
 
     static PrepareImageComponent = (
         component: Component,
-        images: ImageFile[],
+        images: ImageData[],
         pageName: string): JSX.Element | null => {
         return (
             images.length > 1 ? (